requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\HP\Desktop\Electron\Jarvis\recognizer.py: 2
Pillow == 8.1.0

# C:\Users\HP\Desktop\Electron\Jarvis\backend.py: 3
SpeechRecognition == 3.8.1

# C:\Users\HP\Desktop\Electron\Jarvis\backend.py: 10,11
beautifulsoup4 == 4.9.3

# C:\Users\HP\Desktop\Electron\Jarvis\backend.py: 9
google == 3.0.0

pyobjc-framework-cocoa == 7.3

pyttsx3 == 2.90

requests == 2.25.1

setuptools == 47.1.0

torch == 1.8.0

torchvision == 0.10.0

wikipedia == 1.4.0
